requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
astroid==2.15.6
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
azure-cognitiveservices-speech==1.31.0
backcall==0.2.0
blobfile==2.0.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
dataclasses-json==0.5.14
decorator==5.1.1
dill==0.3.7
docopt==0.6.2
elevenlabs==0.2.24
executing==1.2.0
filelock==3.12.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
install==1.3.5
ipython==8.14.0
isort==5.12.0
jedi==0.19.0
langchain==0.0.268
langsmith==0.0.25
lazy-object-proxy==1.9.0
lxml==4.9.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg==3.1.10
psycopg-binary==3.1.10
pure-eval==0.2.2
PyAudio==0.2.13
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.12
pydub==0.25.1
Pygments==2.16.1
pylint==2.17.5
pylint-venv==3.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-osc==1.8.3
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
SQLAlchemy==2.0.20
stack-data==0.6.2
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.4.0
tomlkit==0.12.1
tqdm==4.66.1
traitlets==5.9.0
twitchAPI==3.11.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
webrtcvad==2.0.10
websockets==11.0.3
wrapt==1.15.0
yarg==0.1.9
yarl==1.9.2
